feat(expenses): BUN-25 - Refactor CreateExpenseModal and enhance user experience
- Replaced CreatePageDrawer with CreateExpenseModal for a more integrated expense creation experience.
- Implemented loading state with CreatePageSkeleton while fetching user data.
- Updated CreateExpensePage to directly use user data from getCurrentUser.
- Removed the deprecated CreatePageDrawer component to streamline the codebase.
